requirements.txt
ggDjango==3.0.7
django-cors-headers==3.3.0
six==1.15.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2020.1
psycopg2==2.8.4
gunicorn
django-heroku

